Added partial classes for SoulsDLCPlayer
CrossplayerCommon -> calls methods for other partial classes
Crossplayer{Mod} -> methods / stuff for specific mods

Also clarified some method uses in DLCItem
